some more updates

+++ src/app/(backend)/app/request-items/columns.js
@@ -33,39 +33,47 @@ export const columns = ({ onEdit, onDelete }) => [
     header: "Amount",
   },
   {
+    accessorKey: "remarks",
+    header: "Remarks",
+  },
+  {
     id: "actions",
     cell: ({ row }) => {
       const requestItem = row.original;
 
-      return (
-        <div className="flex justify-end">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="ghost" className="h-8 w-8 p-0">
-                <span className="sr-only">Open menu</span>
-                <MoreHorizontal className="h-4 w-4" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end">
-              <DropdownMenuLabel>Actions</DropdownMenuLabel>
-              <DropdownMenuItem
-                onClick={() => {
-                  onEdit(requestItem);
-                }}
-              >
-                Edit Item
-              </DropdownMenuItem>
-              <DropdownMenuItem
-                onClick={() => {
-                  onDelete(requestItem);
-                }}
-              >
-                Delete Item
-              </DropdownMenuItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-        </div>
-      );
+      if (onEdit && onDelete) {
+        return (
+          <div className="flex justify-end">
+            <DropdownMenu>
+              <DropdownMenuTrigger asChild>
+                <Button variant="ghost" className="h-8 w-8 p-0">
+                  <span className="sr-only">Open menu</span>
+                  <MoreHorizontal className="h-4 w-4" />
+                </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent align="end">
+                <DropdownMenuLabel>Actions</DropdownMenuLabel>
+                <DropdownMenuItem
+                  onClick={() => {
+                    onEdit(requestItem);
+                  }}
+                >
+                  Edit Item
+                </DropdownMenuItem>
+                <DropdownMenuItem
+                  onClick={() => {
+                    onDelete(requestItem);
+                  }}
+                >
+                  Delete Item
+                </DropdownMenuItem>
+              </DropdownMenuContent>
+            </DropdownMenu>
+          </div>
+        );
+      }
+
+      return null;
     },
   },
 ];
